+++ src/main/java/it/gov/pagopa/pu/workflow/wf/pagopa/paymentsreporting/config/BrokersPaymentsReportingPagoPaFetchWfConfig.java
@@ -3,8 +3,8 @@ package it.gov.pagopa.pu.workflow.wf.pagopa.paymentsreporting.config;
 import io.temporal.workflow.Workflow;
 import it.gov.pagopa.payhub.activities.activity.organization.BrokersRetrieverActivity;
 import it.gov.pagopa.payhub.activities.activity.organization.OrganizationBrokeredActiveRetrieverActivity;
-import it.gov.pagopa.pu.workflow.config.BaseWfConfig;
-import it.gov.pagopa.pu.workflow.config.TemporalWFImplementationCustomizer;
+import it.gov.pagopa.pu.workflow.config.temporal.BaseWfConfig;
+import it.gov.pagopa.pu.workflow.config.temporal.TemporalWFImplementationCustomizer;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
