+++ src/main/java/it/gov/pagopa/pu/workflow/wf/pagopa/taxonomy/config/SynchronizeTaxonomyPagoPaFetchWfConfig.java
@@ -2,8 +2,8 @@ package it.gov.pagopa.pu.workflow.wf.pagopa.taxonomy.config;
 
 import io.temporal.workflow.Workflow;
 import it.gov.pagopa.payhub.activities.activity.taxonomy.SynchronizeTaxonomyActivity;
-import it.gov.pagopa.pu.workflow.config.BaseWfConfig;
-import it.gov.pagopa.pu.workflow.config.TemporalWFImplementationCustomizer;
+import it.gov.pagopa.pu.workflow.config.temporal.BaseWfConfig;
+import it.gov.pagopa.pu.workflow.config.temporal.TemporalWFImplementationCustomizer;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
