+++ src/main/java/it/gov/pagopa/pu/workflow/wf/pagopa/taxonomy/wftaxonomyfetch/SynchronizeTaxonomyPagoPaFetchWFImpl.java
@@ -3,6 +3,7 @@ package it.gov.pagopa.pu.workflow.wf.pagopa.taxonomy.wftaxonomyfetch;
 
 import io.temporal.spring.boot.WorkflowImpl;
 import it.gov.pagopa.payhub.activities.activity.taxonomy.SynchronizeTaxonomyActivity;
+import it.gov.pagopa.pu.workflow.config.temporal.TemporalWFImplementationCustomizer;
 import it.gov.pagopa.pu.workflow.wf.pagopa.taxonomy.config.SynchronizeTaxonomyPagoPaFetchWfConfig;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeansException;
@@ -20,7 +21,7 @@ public class SynchronizeTaxonomyPagoPaFetchWFImpl implements SynchronizeTaxonomy
 /**
    * Temporal workflow will not allow to use injection in order to avoid <a href="https://docs.temporal.io/workflows#non-deterministic-change">non-deterministic changes</a> due to dynamic reconfiguration.<BR />
    * Anyway it allows to override ActivityOptions, but actually it's not supporting the override based on the particular workflow.<BR />
-   * In {@link it.gov.pagopa.pu.workflow.config.TemporalWFImplementationCustomizer} we are already setting defaults to all workflows.<BR />
+   * In {@link TemporalWFImplementationCustomizer} we are already setting defaults to all workflows.<BR />
    * Use this as an example to override based on the particular workflow.
    */
